Cambio en el tipo de dato del kilometraje, de decimal a entero, esto cambia la logica de varios metodos de algunos controladores y partes de las vistas

+++ PruebaNETCore/Models/FichaMantencion.cs
@@ -9,7 +9,7 @@ public partial class FichaMantencion
 
     public DateTime FechaMantencion { get; set; }
 
-    public decimal Kilometraje { get; set; }
+    public int Kilometraje { get; set; }
 
     public string Descripcion { get; set; }
